better

+++ src/main/resources/db/migration/V1.2__item_added.sql
@@ -2,7 +2,7 @@ create table item(
     id int primary key,
     name varchar(50),
     length int,
-    weight numeric(10,5),
+    weight float,
     registration timestamp,
     fake boolean,
     state varchar(50)
